Mismatched types: j was infered as 'float64' (because of the underlying platform)

+++ 03-types/main.go
@@ -3,8 +3,8 @@ package main
 import "fmt"
 
 func main() {
-	i := 3            // int
-	j := 2.3          // float
-	sum := i + int(j) // Explicit conversion to type 'int'
-	fmt.Println(sum)  // prints 5
+	i := 3                // int
+	j := 2.3              // float
+	sum := float32(i) + j // ERROR: infered type for j is 'float64'!!!
+	fmt.Println(sum)
 }
